SASI (autoplanning)
The script was made with a binary file to which a user control window was added with which the user can choose between different processes through a combobox among the processes are the creation of structures such as high-density artifacts (HDA) greater than 3.0 g /cc, the ring structure which has a thickness of 1.0 cm to 2.0 cm from the Planning Target Volume (PTV) and creation of the Optimization Target Volume (OTV) structure, the creation of a treatment plan, creation of the Computed Cone Beam Tomography (CBCT) field, in addition to allowing the review of the names of the structures that are evaluated in the clinical system. This script is the result of many various ideas found in forums and they are summarized for the clinic where I work. I hope it will help you automate processes and adapt it just like I did